requirements.txt
python-binance
matplotlib
asyncio
